Cover VS (desktop MSBuild) in end-to-end tests
Nothing exercised the .NET Framework MSBuild.exe that Visual Studio
loads the task under; CI only ran Core MSBuild via dotnet. Move the
build-driving tests into their own project (Publicizer.E2ETests) and
run it in a matrix, selecting the host per leg (dotnet on Linux,
desktop msbuild on Windows) via PUBLICIZER_TEST_BUILDER. Unit tests
now run Linux-only.

Trim to two cases (Publicize, PublicizeAll) that prove the MSBuild
wiring; per-member behavior lives in the characterization and engine
tests.
