+++ src/Publicizer.E2ETests/ProcessResult.cs
@@ -1,3 +1,3 @@
-namespace Publicizer.Tests;
+namespace Publicizer.E2ETests;
 
 internal sealed record ProcessResult(int ExitCode, string Output, string Error);
