+++ src/Publicizer.Tests/Publicizer.Tests.csproj
@@ -17,13 +17,4 @@
     <ProjectReference Include="..\Publicizer\Publicizer.csproj" />
   </ItemGroup>
 
-  <Target Name="CopyNugetPackage" AfterTargets="AfterBuild">
-    <ItemGroup>
-      <PublicizerNuget Include="..\Publicizer\bin\Krafs.Publicizer.*.nupkg"  />
-      <StaleNuget Include="$(OutputPath)Krafs.Publicizer.*.nupkg" />
-    </ItemGroup>
-    <Delete Files="@(StaleNuget)" />
-    <Copy SourceFiles="@(PublicizerNuget)" DestinationFolder="$(OutputPath)" />
-  </Target>
-
 </Project>
